pyproject.toml
[tool.poetry]
name = "BiscoteGirl"
version = "0.2.0"
description = "Bot de réservation fitness avec notifications enrichies"
authors = ["gsoulat31@gmail.com"]
[tool.poetry.dependencies]
python = "^3.12"
beautifulsoup4 = "^4.12.3"
playwright = "^1.49.1"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
loguru = "^0.7.2"
tenacity = "^9.0.0"
tqdm = "^4.67.1"
python-dateutil = "^2.9.0.post0"
ruff = "^0.9.0"
pillow = "^11.1.0"
fastapi = "^0.115.6"
uvicorn = "^0.34.0"
pydantic = "^2.10.5"
[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
black = "^24.1.1"
flake8 = "^7.0.0"
mypy = "^1.8.0"
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"